edited jarsigner command at :apksignerkey
in jre 1.7, you need to add -sigalg and -digestalg to the jarsigner command. if not, the APK will still sign, and google play will accept it but *it won't install on users devices*!

i also removed the %1 at the end of the command, i can't see what parameter you're using.